refactor: widget 폴더 생성 후 헤더 파일 이동 및 widget 별칭 설정
