+++ frontend/.eslintrc.json
@@ -42,6 +42,11 @@
             "position": "before"
           },
           {
+            "pattern": "@widgets/**",
+            "group": "internal",
+            "position": "before"
+          },
+          {
             "pattern": "@components/**",
             "group": "internal",
             "position": "before"
