+++ frontend/src/pages/CreateRoom.tsx
@@ -3,7 +3,7 @@ import InviteMember from '@domains/room/components/RoomDetailTab/InviteMember';
 import Button from '@components/actions/Button';
 import Input from '@components/actions/Input';
 import ErrorMessage from '@components/errors/ErrorMessage';
-import { HEADER_HEIGHT } from '@components/layouts/Header';
+import { HEADER_HEIGHT } from 'widgets/Header';
 
 import { useCreateRoom } from '@domains/room/hooks/useCreateRoom';
 import { useInviteMember } from '@domains/room/hooks/useInviteMember';
