+++ frontend/src/pages/ProfileInit.tsx
@@ -1,6 +1,6 @@
 import Button from '@components/actions/Button';
 import Input from '@components/actions/Input';
-import { HEADER_HEIGHT } from '@components/layouts/Header';
+import { HEADER_HEIGHT } from 'widgets/Header';
 
 import { useAuth } from '@domains/login/context/AuthProvider';
