+++ frontend/src/shared/provider/ToastProvider.tsx
@@ -1,4 +1,4 @@
-import { HEADER_HEIGHT } from '@components/layouts/Header';
+import { HEADER_HEIGHT } from 'widgets/Header';
 import Toast, { DEFAULT_TIME } from '@components/toast/Toast';
 
 import { generateRandomNumber } from '@utils/generateRandomNumber';
